4.0 dev merge release 4.1.x

See merge request dt-insight-engine/DAGScheduleX!597

+++ engine-master/engine-master-server/src/main/java/com/dtstack/engine/master/jobdealer/JobCheckpointDealer.java
@@ -121,7 +121,7 @@ public class JobCheckpointDealer implements InitializingBean {
                     JobCheckpointInfo taskInfo = delayBlockingQueue.take();
                     String engineJobId = taskInfo.getJobIdentifier().getEngineJobId();
                     ScheduleJob jobInfo = scheduleJobDao.getByJobId(taskInfo.getTaskId(), 0);
-                    int status = jobInfo.getStatus().intValue();
+                    int status = jobInfo.getStatus();
 
                     updateCheckpointImmediately(taskInfo, engineJobId, status);
 
@@ -254,7 +254,7 @@ public class JobCheckpointDealer implements InitializingBean {
      * @return
      * @throws ExecutionException
      */
-    private Map getJobParamsByJobId(String jobId) throws ExecutionException {
+    private Map getJobParamsByJobId(String jobId) throws ExecutionException{
         Map<String, Object> taskParams = checkpointConfigCache.get(jobId, () -> {
             Map<String, Object> paramInfo = Maps.newConcurrentMap();
             String jobInfo = engineJobCacheDao.getOne(jobId).getJobInfo();
